Move table tab panels to its own directory
