+++ packages/extension/src/view/devtools/components/privateAdvertising/protectedAudience/explorableExplanation/tableTabPanels/infoPanel.tsx
